% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{lu_decompose}
\alias{lu_decompose}
\title{LU decomposition of a square matrix}
\usage{
lu_decompose(A)
}
\arguments{
\item{A}{A square numeric matrix.}
}
\value{
A list containing:
\itemize{
  \item \code{L} - lower triangular matrix,
  \item \code{U} - upper triangular matrix,
  \item \code{P} - permutation matrix,
  \item \code{swaps} - number of row swaps.
}
}
\description{
The function performs an LU decomposition of a matrix with partial pivoting 
using C code with OpenMP.
}
\examples{
A <- matrix(c(2, 4, -2, 1, -6, 7, 1, 0, 2), nrow = 3)
lu_decompose(A)

}
