% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{lu_det}
\alias{lu_det}
\title{Matrix determinant calculation using LU decomposition}
\usage{
lu_det(A)
}
\arguments{
\item{A}{A square numeric matrix.}
}
\value{
The numeric value of the determinant. If the matrix is singular, it returns 0.
}
\description{
The function calculates the determinant of a square matrix based on its LU decomposition.
}
\examples{
A <- matrix(c(2, 4, -2, 1, -6, 7, 1, 0, 2), nrow = 3)
lu_det(A)

}
