% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{sor_solve}
\alias{sor_solve}
\title{Solving systems of linear equations using SOR method}
\usage{
sor_solve(A, b, omega = 1, tol = 1e-07, max_iter = 1000L)
}
\arguments{
\item{A}{A square numeric matrix of the system's coefficients.}

\item{b}{A numeric vector of the right-hand side constants.}

\item{omega}{The relaxation parameter. For \code{omega = 1}, the algorithm reduces 
to the Gauss-Seidel method. Default is 1.0.}

\item{tol}{Error tolerance defining the stopping criterion. Default is 1e-7.}

\item{max_iter}{Maximum number of iterations. Default is 1000.}
}
\value{
A list containing:
\itemize{
  \item \code{x} - a vector representing the approximate solution of the system,
  \item \code{iterations} - the number of iterations performed.
}
}
\description{
The iterative Successive Over-Relaxation (SOR) method for solving
systems of linear equations.
}
\examples{
A <- matrix(c(10, -1, 2, 0, -1, 11, -1, 3, 2, -1, 10, -1, 0, 3, -1, 8), 4, 4, byrow=TRUE)
b <- c(6, 25, -11, 15)
sor_solve(A, b, omega = 1.1)

}
