% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{lu_solve}
\alias{lu_solve}
\title{Solving systems of linear equations using LU decomposition}
\usage{
lu_solve(A, b)
}
\arguments{
\item{A}{A square numeric matrix of the system's coefficients.}

\item{b}{A numeric vector of the right-hand side constants.}
}
\value{
A numeric vector representing the solution to the system of equations.
}
\description{
The function solves a system of linear equations \code{Ax = b} using 
LU decomposition.
}
\examples{
A <- matrix(c(2, 4, -2, 1, -6, 7, 1, 0, 2), nrow = 3)
b <- c(4, -2, 7)
lu_solve(A, b)

}
